Refact organization profile edit page
