+++ src/app/(auth)/profile/edit/components/ProfileEditMobileView.tsx
@@ -6,7 +6,7 @@ import { useApp } from "@/contexts/AppContext";
 import { useAvatars } from "@/hooks/useAvatars";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import CapacitySelectionModal from "./CapacitySelectionModal";
+import CapacitySelectionModal from "@/components/CapacitySelectionModal";
 import AccountCircleIcon from "@/public/static/images/account_circle.svg";
 import AccountCircleIconWhite from "@/public/static/images/account_circle_white.svg";
 import AccountBoxIcon from "@/public/static/images/account_box.svg";
